+++ src/Phuria/SQLBuilder/Connection/ConnectionManager.php
@@ -1,0 +1,39 @@
+<?php
+
+/**
+ * This file is part of Phuria SQL Builder package.
+ *
+ * Copyright (c) 2016 Beniamin Jonatan Šimko
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Phuria\SQLBuilder\Connection;
+
+/**
+ * @author Beniamin Jonatan Šimko <spam@simko.it>
+ */
+class ConnectionManager implements ConnectionManagerInterface
+{
+    /**
+     * @var array
+     */
+    private $connections = [];
+
+    /**
+     * @inheritdoc
+     */
+    public function registerConnection(ConnectionInterface $connection, $name = 'default')
+    {
+        $this->connections[$name] = $connection;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getConnection($name = null)
+    {
+        return $name ? $this->connections[$name] : reset($this->connections);
+    }
+}