+++ src/Phuria/SQLBuilder/Statement/NullStatement.php
@@ -1,0 +1,50 @@
+<?php
+
+/**
+ * This file is part of Phuria SQL Builder package.
+ *
+ * Copyright (c) 2016 Beniamin Jonatan Šimko
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Phuria\SQLBuilder\Statement;
+
+/**
+ * @author Beniamin Jonatan Šimko <spam@simko.it>
+ */
+class NullStatement implements StatementInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function bindValue($parameter, $value)
+    {
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function fetchScalar()
+    {
+        return null;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function execute()
+    {
+        return $this;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function rowCount()
+    {
+        return null;
+    }
+}