+++ tests/Phuria/SQLBuilder/Test/Fixtures/NullQueryBuilder.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * This file is part of Phuria SQL Builder package.
+ *
+ * Copyright (c) 2016 Beniamin Jonatan Šimko
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Phuria\SQLBuilder\Test\Fixtures;
+
+use Phuria\SQLBuilder\QueryBuilder\AbstractBuilder;
+
+/**
+ * @author Beniamin Jonatan Šimko <spam@simko.it>
+ */
+class NullQueryBuilder extends AbstractBuilder
+{
+    public function __construct() {}
+}