+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
++ ConnectionManager to handling more that one connection
++ NullConnection, NullStatement classes
 ### Changed
 + QueryBuilder class name to QueryBuilderFactory, and factory methods
 now will be prefixed 'create'